SONAR-7287 Move FavoriteService to favorite package
