+++ server/sonar-server/src/main/java/org/sonar/server/favorite/ws/AddAction.java
@@ -29,7 +29,7 @@ import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.component.ComponentDto;
 import org.sonar.server.component.ComponentFinder;
-import org.sonar.server.component.FavoriteService;
+import org.sonar.server.favorite.FavoriteService;
 import org.sonar.server.user.UserSession;
 import org.sonar.server.ws.KeyExamples;
