+++ server/sonar-server/src/test/java/org/sonar/server/favorite/ws/FavoritesWsTest.java
@@ -24,7 +24,7 @@ import org.junit.Test;
 import org.sonar.api.server.ws.WebService.Controller;
 import org.sonar.db.DbClient;
 import org.sonar.server.component.ComponentFinder;
-import org.sonar.server.component.FavoriteService;
+import org.sonar.server.favorite.FavoriteService;
 import org.sonar.server.user.UserSession;
 import org.sonar.server.ws.WsTester;
